[module7-task3]: Extract auth.router from user.router
