+++ src/service/server/routes/api.ts
@@ -4,6 +4,7 @@ import {articleRouter} from "./articles.router";
 import {categoriesRouter} from "./categories.router";
 import {
   articlesControllerFactory,
+  authControllerFactory,
   categoriesControllerFactory,
   commentsControllerFactory,
   searchControllerFactory,
@@ -16,6 +17,7 @@ import {ICategoryModel} from "../data-access/models/category";
 import {IArticleModel} from "../data-access/models/article";
 import {ICommentModel} from "../data-access/models/comment";
 import {IUserModel} from "../data-access/models/user";
+import {authRouter} from "./auth.router";
 
 export const apiRouter = ({
   CategoryModel,
@@ -38,12 +40,14 @@ export const apiRouter = ({
   const searchController = searchControllerFactory({ArticleModel});
   const commentsController = commentsControllerFactory({CommentModel});
   const usersController = usersControllerFactory({UserModel});
+  const authController = authControllerFactory({UserModel});
 
   router.use(APIRoutes.ARTICLES, articleRouter(articlesController, commentsController));
   router.use(APIRoutes.CATEGORIES, categoriesRouter(articlesController, categoriesController));
   router.use(APIRoutes.CATEGORIES_STATISTICS, categoriesStatisticsRouter(categoriesController));
   router.use(APIRoutes.SEARCH, searchRouter(searchController));
   router.use(APIRoutes.USERS, usersRouter(usersController));
+  router.use(APIRoutes.AUTH, authRouter(authController));
 
   return router;
 };
