+++ src/service/server/routes/users.router.e2e.test.ts
@@ -6,24 +6,9 @@ import {initApp} from "./tests-boilerplate/init-app";
 import {UserId} from "../../../types/user-id";
 import {IUserPreview} from "../../../types/interfaces/user-preview";
 import {IUserCreatingDoublePasswords} from "../../../types/interfaces/user-creating";
-import {ILogin} from "../../../types/interfaces/login";
-import {IAuthorizationSuccess} from "../../../types/interfaces/authorization-result";
-import {IAuthTokens} from "../../../types/interfaces/auth-tokens";
 
 const validUserId: UserId = 2;
 const invalidUserId = `-1`;
-const validSignUp: ILogin = {
-  email: `zaberkirder8@usgs.gov`,
-  password: `hzgdghdglhdgklgz`,
-};
-const validUserToSignUp: IUserCreatingDoublePasswords = {
-  ...validSignUp,
-  firstName: `Lowe`,
-  lastName: `Tennant`,
-  avatar: ``,
-  roleId: 2,
-  passwordRepeated: `hzgdghdglhdgklgz`,
-};
 const validNewUser: IUserCreatingDoublePasswords = {
   email: `zaberkirder8@usgs.gov`,
   firstName: `Lowe`,
@@ -96,44 +81,6 @@ describe(`Users router`, () => {
       expect(responseKeys).toContain(`roleId`);
       expect(responseKeys).toContain(`password`);
       expect(responseKeys).toContain(`passwordRepeated`);
-    });
-  });
-
-  describe(`/login`, () => {
-    beforeAll(async () => await request(app).post(`/api/users/`).send(validUserToSignUp));
-
-    describe(`POST user`, () => {
-      test(`Should return code 200 when sign-up successfully`, async () => {
-        const res = await request(app).post(`/api/users/login`).send(validSignUp);
-        expect(res.status).toBe(200);
-      });
-      test(`Should return access token and refresh token when sign-up successfully`, async () => {
-        const res = await request(app).post(`/api/users/login`).send(validSignUp);
-        const responseKeys = Object.keys(res.body as IAuthorizationSuccess);
-        const payloadKeys = Object.keys((res.body as IAuthorizationSuccess).payload as IAuthTokens);
-        expect(responseKeys).toContain(`payload`);
-        expect(responseKeys).toContain(`isSuccess`);
-        expect(payloadKeys).toContain(`accessToken`);
-        expect(payloadKeys).toContain(`refreshToken`);
-      });
-      test(`Should return code 400 when pass invalid password`, async () => {
-        const res = await request(app)
-          .post(`/api/users/login`)
-          .send({
-            ...validSignUp,
-            password: `invalid-password`,
-          });
-        expect(res.status).toBe(403);
-      });
-      test(`Should return code 400 when pass non-existing email`, async () => {
-        const res = await request(app)
-          .post(`/api/users/login`)
-          .send({
-            ...validSignUp,
-            email: `non-existing-email@gmail.com`,
-          });
-        expect(res.status).toBe(403);
-      });
     });
   });
 });
