+++ src/service/server/routes/users.router.ts
@@ -2,7 +2,6 @@ import {Router} from "express";
 import {UsersController} from "../controllers/users.controller";
 import {HttpCode} from "../../../constants-es6";
 import {validateNewUser} from "../validators";
-import {validateLogin} from "../validators/validate-login";
 
 export const usersRouter = (usersController: UsersController): Router => {
   const router = Router();
@@ -20,16 +19,6 @@ export const usersRouter = (usersController: UsersController): Router => {
       return res.status(status).send(payload);
     } catch (e) {
       return res.status(HttpCode.BAD_REQUEST).send(e);
-    }
-  });
-
-  router.post(`/login`, async (req, res) => {
-    try {
-      const {email, password} = await validateLogin(req.body);
-      const {status = HttpCode.OK, payload} = await usersController.login({email, password});
-      return res.status(status).send(payload);
-    } catch (e) {
-      return res.status(HttpCode.FORBIDDEN).send(e);
     }
   });
